requirements.txt
asgiref==3.7.2
blinker==1.7.0
boto3==1.34.76
botocore==1.34.76
cachelib==0.10.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
cs50==9.3.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-haystack==3.2.1
django-resized==1.0.2
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-extra-fields==3.7.0
drf_base64==2.0
filetype==1.2.0
Flask==3.0.0
Flask-Session==0.5.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
greenlet==3.0.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.5.6
MarkupSafe==2.1.3
oauth2client==4.1.3
packaging==23.2
phonenumbers==8.13.29
pillow==10.2.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.11.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2023.3.post1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.10.1
setuptools==69.2.0
six==1.16.0
sorl-thumbnail==12.10.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
termcolor==2.3.0
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
wheel==0.41.3
whitenoise==6.6.0
